+++ src/main/java/br/edu/ifto/pwebII/controller/PacienteController.java
@@ -1,9 +1,12 @@
 package br.edu.ifto.pwebII.controller;
 
+import br.edu.ifto.pwebII.model.entity.Consulta;
 import br.edu.ifto.pwebII.model.entity.Paciente;
+import br.edu.ifto.pwebII.model.jdbc.repository.ConsultaRepository;
 import br.edu.ifto.pwebII.model.jdbc.repository.PacienteRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
@@ -11,17 +14,21 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import java.util.List;
+
 @Transactional
 @Controller
 @RequestMapping("paciente")
 public class PacienteController {
 
     private final PacienteRepository repository;
+    private final ConsultaRepository consultaRep;
 
     // O Spring injetará o repositório automaticamente
     @Autowired
-    public PacienteController(PacienteRepository repository) {
+    public PacienteController(PacienteRepository repository, ConsultaRepository consultaRep) {
         this.repository = repository;
+        this.consultaRep = consultaRep;
     }
 
     @GetMapping("/form")
@@ -42,15 +49,17 @@ public class PacienteController {
     }
 
     @PostMapping("/remove/{id}")
-    public String remove(@PathVariable Long id, RedirectAttributes redirectAttributes) {
-        boolean removido = repository.remove(id);
+    public String remove(@PathVariable("id") Long id, RedirectAttributes redirectAttributes) {
+        List<Consulta> consultasDoPaciente = consultaRep.consultasPorPaciente(id);
 
-        if (removido) {
-            redirectAttributes.addFlashAttribute("mensagemSucesso", "Paciente excluído com sucesso!");
-        } else {
+        if (!consultasDoPaciente.isEmpty()){
             redirectAttributes.addFlashAttribute("mensagemErro",
-                    "Não é possível excluir o paciente.");
+                    "Não é possível excluir o(a) paciente " + repository.paciente(id).getNome() + ", pois ele(a) possui consultas vinculadas a ele(a).");
+            return "redirect:/paciente/list";
         }
+
+        repository.remove(id);
+        redirectAttributes.addFlashAttribute("mensagemSucesso", "Paciente excluído com sucesso!");
 
         return "redirect:/paciente/list";
     }
